HLSStreamWorker: avoid a possible integer overflow
Don't think retries will ever get big enough for it to be a problem in
reality. Fixes Coverity ID 1192755 Unintentional integer overflow

+++ mythtv/libs/libmythtv/recorders/HLS/HLSStreamWorker.cpp
@@ -71,7 +71,7 @@ void HLSStreamWorker::run(void)
             if (retries == 10)
                 m_parent->ResetSequence();
 
-            delay = 500 * retries * retries;
+            delay = (uint64_t)500 * retries * retries;
             if (delay > 20000)
                 delay = 20000;
         }
